requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
psycopg2-binary==2.9.1
python-dotenv==0.19.0
requests==2.26.0
gunicorn==20.1.0
Werkzeug==2.0.1
python-jose==3.3.0